feat: take final screenshot
