+++ src/lib/constants/stickerConfig.ts
@@ -1,3 +1,5 @@
 import { StickerMenu } from '@/types'
 
 export const STICKER_MENU: StickerMenu[] = [0, 1, 2, 3]
+
+export const GUEST_STICKER_MENU: StickerMenu[] = [1, 2, 3]
